+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - #12 - Order entity
 - #61 - Front CRUD for Passport
 - #92 - Security service
+- #101 - added fuzzy search for product
 
 ## [1.1.0] - 8.09.2023
 - #5 - added user entity